merge branch integration_2026-06-04_1161487139074 into master

See merge request: !533

+++ src/Privatezone/Model/RecordForListRecordsOutput.php
@@ -61,7 +61,7 @@ class RecordForListRecordsOutput implements ModelInterface, ArrayAccess
         'updated_at' => null,
         'value' => null,
         'weight' => 'int32',
-        'zid' => 'int32'
+        'zid' => 'int64'
     ];
 
     /**
